feat: add EventBus & autoComplete for input / update SignInForm & AuthControl

+++ src/components/authControl/authControl.tsx
@@ -1,3 +1,5 @@
+'use client ';
+
 import { FC, useEffect, useState } from 'react';
 import Button from '@mui/material/Button';
 import { useRouter } from 'next/navigation';
@@ -5,40 +7,57 @@ import { ROUTES } from '@/utils/constants/routes.ts';
 import { logout } from '@/utils/firebase/firebase';
 import { deleteCookie } from '@/utils/cookies/deleteCookie';
 import { getCookie } from '@/utils/cookies/getCookie';
-
 import styles from './authControl.module.scss';
+import { EUserEvent, localEventBus } from '@/utils/EventBus';
 
 const AuthControl: FC = (): JSX.Element => {
-  // const [isAuth, setIsAuth] = useState<boolean | null>(null);
-  const [isAuth, setIsAuth] = useState<boolean>(!!getCookie('authToken'));
+  const [isAuth, setIsAuth] = useState<boolean>(false);
+  const [loading, setLoading] = useState<boolean>(true);
   const router = useRouter();
 
-  // useEffect((): void => {
-  // const authToken = getCookie('authToken');
-  //   // setIsAuth(!!authToken);
-  // }, [isAuth]);
-  useEffect((): void => {
-    const authToken = getCookie('authToken');
-    setIsAuth(!!authToken);
+  useEffect((): (() => void) => {
+    const { exists } = getCookie('authToken');
+    setIsAuth(exists);
+
+    const unsubLogin = localEventBus.subscribeToEvent(
+      EUserEvent.USER_LOGIN,
+      (): void => {
+        setIsAuth(true);
+      }
+    );
+    const unsubLogout = localEventBus.subscribeToEvent(
+      EUserEvent.USER_LOGOUT,
+      (): void => {
+        setIsAuth(false);
+      }
+    );
+
+    setLoading(false);
+
+    return (): void => {
+      unsubLogin();
+      unsubLogout();
+    };
   }, []);
 
-  const handleSignOut = async () => {
+  const handleSignOut = async (): Promise<void> => {
     try {
-      logout();
-      deleteCookie('authToken');
-      setIsAuth(false);
-      router.push(ROUTES.HOME_PATH);
+      const logoutUserAuth = await logout();
+      if (!logoutUserAuth.userAuth) {
+        await deleteCookie('authToken');
+        localEventBus.emitEvent(EUserEvent.USER_LOGOUT);
+        router.push(ROUTES.HOME_PATH);
+      }
     } catch (error) {
       console.error('Error signing out:', error);
     }
   };
 
-  const handleSignIn = () => {
-    router.push(ROUTES.SIGN_IN_PATH);
-  };
+  const handleSignIn = (): void => router.push(ROUTES.SIGN_IN_PATH);
+  const handleSignUp = (): void => router.push(ROUTES.SIGN_UP_PATH);
 
-  function signUp() {
-    router.push(ROUTES.SIGN_UP_PATH);
+  if (loading) {
+    return <></>;
   }
 
   return isAuth ? (
@@ -50,7 +69,7 @@ const AuthControl: FC = (): JSX.Element => {
       <Button variant="contained" onClick={handleSignIn}>
         Sign in
       </Button>
-      <Button variant="contained" onClick={signUp}>
+      <Button variant="contained" onClick={handleSignUp}>
         Sign Up
       </Button>
     </div>
